+++ prisma/seeds/social.mjs
@@ -4,19 +4,24 @@ import logger from 'node-color-log';
 async function seed() {
   const count = await prisma?.social?.count();
 
-  if (count && count > 0) {
-    await prisma?.social?.deleteMany();
-  }
+  if (count && count > 0) await prisma?.social?.deleteMany();
 
-  await prisma?.social?.createMany({
-    data: [],
-  });
+  const data = [
+    {
+      name: 'Github',
+      url: 'https://github.com/mflisikowski/',
+    },
+    {
+      name: 'Linkedin',
+      url: 'https://www.linkedin.com/in/mflisikowski/',
+    },
+  ];
+
+  try {
+    await prisma?.social?.createMany({ data });
+  } catch (error) {
+    logger.error(error.message);
+  }
 }
 
-// name   String
-// url    String
-// mailto String?
-
-seed()
-  .catch((e) => logger.error(e) && process.exit(1))
-  .finally(async () => await prisma.$disconnect());
+seed().finally(async () => await prisma.$disconnect());
